backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
torchvision
pillow
pytesseract
python-multipart